[refactor]: Auth 로직을 다른 파일로 분리 (#26)
